requirements-dev.txt
numpy
pybullet
-r requirements-optional.txt

# dev dependencies
pycddlib
scipy
